test: refactoring package name to data
