+++ jnosql-communication/jnosql-communication-query/src/test/java/org/eclipse/jnosql/communication/query/data/DeleteJakartaDataQueryProviderAndTest.java
@@ -1,15 +1,15 @@
 /*
  *  Copyright (c) 2024 Contributors to the Eclipse Foundation
- *  All rights reserved. This program OR the accompanying materials
+ *  All rights reserved. This program and the accompanying materials
  *  are made available under the terms of the Eclipse Public License v1.0
- *  OR Apache License v2.0 which accompanies this distribution.
+ *  and Apache License v2.0 which accompanies this distribution.
  *  The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *  OR the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
+ *  and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
  *  You may elect to redistribute this code under either of these licenses.
  *  Contributors:
  *  Otavio Santana
  */
-package org.eclipse.jnosql.communication.query.data.delete;
+package org.eclipse.jnosql.communication.query.data;
 
 import org.assertj.core.api.SoftAssertions;
 import org.eclipse.jnosql.communication.Condition;
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class DeleteJakartaDataQueryProviderOrTest {
+class DeleteJakartaDataQueryProviderAndTest {
 
 
     private DeleteProvider deleteProvider;
@@ -34,8 +34,8 @@ class DeleteJakartaDataQueryProviderOrTest {
 
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = "DELETE FROM entity WHERE age = 10 OR salary = 10.15")
-    void shouldORTwoConditions(String query){
+    @ValueSource(strings = {"DELETE FROM entity WHERE age = 10 AND salary = 10.15"})
+    void shouldAndTwoConditions(String query){
         var deleteQuery = deleteProvider.apply(query);
 
         SoftAssertions.assertSoftly(soft -> {
@@ -44,7 +44,7 @@ class DeleteJakartaDataQueryProviderOrTest {
             soft.assertThat(deleteQuery.where()).isNotEmpty();
             var where = deleteQuery.where().orElseThrow();
             var condition = where.condition();
-            soft.assertThat(condition.condition()).isEqualTo(Condition.OR);
+            soft.assertThat(condition.condition()).isEqualTo(Condition.AND);
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
@@ -58,8 +58,8 @@ class DeleteJakartaDataQueryProviderOrTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = "DELETE FROM entity WHERE age = 10 OR salary = 10.15 OR name =?1")
-    void shouldORThreeConditions(String query){
+    @ValueSource(strings = {"DELETE FROM entity WHERE age = 10 AND salary = 10.15 AND name =?1"})
+    void shouldAndThreeConditions(String query){
         var deleteQuery = deleteProvider.apply(query);
 
         SoftAssertions.assertSoftly(soft -> {
@@ -68,7 +68,7 @@ class DeleteJakartaDataQueryProviderOrTest {
             soft.assertThat(deleteQuery.where()).isNotEmpty();
             var where = deleteQuery.where().orElseThrow();
             var condition = where.condition();
-            soft.assertThat(condition.condition()).isEqualTo(Condition.OR);
+            soft.assertThat(condition.condition()).isEqualTo(Condition.AND);
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
@@ -85,7 +85,7 @@ class DeleteJakartaDataQueryProviderOrTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = "DELETE FROM entity WHERE age = 10 OR salary = 10.15 AND name =?1")
+    @ValueSource(strings = {"DELETE FROM entity WHERE age = 10 AND salary = 10.15 OR name =?1"})
     void shouldORMixConditions(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -95,7 +95,7 @@ class DeleteJakartaDataQueryProviderOrTest {
             soft.assertThat(deleteQuery.where()).isNotEmpty();
             var where = deleteQuery.where().orElseThrow();
             var condition = where.condition();
-            soft.assertThat(condition.condition()).isEqualTo(Condition.OR);
+            soft.assertThat(condition.condition()).isEqualTo(Condition.AND);
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
@@ -109,7 +109,7 @@ class DeleteJakartaDataQueryProviderOrTest {
             condition = conditions.get(2);
             values = (ConditionQueryValue) condition.value();
             conditions = values.get();
-            soft.assertThat(condition.condition()).isEqualTo(Condition.AND);
+            soft.assertThat(condition.condition()).isEqualTo(Condition.OR);
             soft.assertThat(conditions).hasSize(1);
 
             soft.assertThat(conditions.get(0).name()).isEqualTo("name");
