+++ jnosql-communication/jnosql-communication-query/src/test/java/org/eclipse/jnosql/communication/query/data/DeleteJakartaDataQueryProviderInTest.java
@@ -9,20 +9,26 @@
  *  Contributors:
  *  Otavio Santana
  */
-package org.eclipse.jnosql.communication.query.data.delete;
+package org.eclipse.jnosql.communication.query.data;
 
 import org.assertj.core.api.SoftAssertions;
 import org.eclipse.jnosql.communication.Condition;
+import org.eclipse.jnosql.communication.query.ArrayQueryValue;
 import org.eclipse.jnosql.communication.query.ConditionQueryValue;
+import org.eclipse.jnosql.communication.query.EnumQueryValue;
 import org.eclipse.jnosql.communication.query.NumberQueryValue;
+import org.eclipse.jnosql.communication.query.QueryCondition;
 import org.eclipse.jnosql.communication.query.QueryValue;
 import org.eclipse.jnosql.communication.query.StringQueryValue;
+import org.eclipse.jnosql.communication.query.data.DefaultQueryValue;
 import org.eclipse.jnosql.communication.query.data.DeleteProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class DeleteJakartaDataQueryProviderLikeTest {
+import java.time.DayOfWeek;
+
+class DeleteJakartaDataQueryProviderInTest {
 
 
     private DeleteProvider deleteProvider;
@@ -33,8 +39,8 @@ class DeleteJakartaDataQueryProviderLikeTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"WHERE  name LIKE 'A%'", "FROM entity WHERE name LIKE 'A%'"})
-    void shouldLike(String query){
+    @ValueSource(strings = {"DELETE FROM entity WHERE age IN (10, 12.12, 'otavio', ?1, :param)"})
+    void shouldIn(String query){
         var deleteQuery = deleteProvider.apply(query);
 
         SoftAssertions.assertSoftly(soft -> {
@@ -44,15 +50,20 @@ class DeleteJakartaDataQueryProviderLikeTest {
             var where = deleteQuery.where().orElseThrow();
             var condition = where.condition();
             QueryValue<?> value = condition.value();
-            soft.assertThat(condition.condition()).isEqualTo(Condition.LIKE);
-            soft.assertThat(value).isEqualTo(StringQueryValue.of("A%"));
+            soft.assertThat(condition.name()).isEqualTo("age");
+            soft.assertThat(condition.condition()).isEqualTo(Condition.IN);
+            soft.assertThat(value).isInstanceOf(ArrayQueryValue.class);
+            soft.assertThat(ArrayQueryValue.class.cast(value).get()).hasSize(5)
+                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(12.12),
+                            StringQueryValue.of("otavio"), DefaultQueryValue.of("?1"),
+                            DefaultQueryValue.of("param"));
 
         });
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"WHERE  name LIKE \"A%\"", "FROM entity WHERE name LIKE \"A%\""})
-    void shouldLikeDoubleQuote(String query){
+    @ValueSource(strings = {"DELETE FROM entity WHERE days IN (java.time.DayOfWeek.MONDAY, java.time.DayOfWeek.SUNDAY)"})
+    void shouldInEnumLiteral(String query){
         var deleteQuery = deleteProvider.apply(query);
 
         SoftAssertions.assertSoftly(soft -> {
@@ -62,15 +73,19 @@ class DeleteJakartaDataQueryProviderLikeTest {
             var where = deleteQuery.where().orElseThrow();
             var condition = where.condition();
             QueryValue<?> value = condition.value();
-            soft.assertThat(condition.condition()).isEqualTo(Condition.LIKE);
-            soft.assertThat(value).isEqualTo(StringQueryValue.of("A%"));
+            soft.assertThat(condition.name()).isEqualTo("days");
+            soft.assertThat(condition.condition()).isEqualTo(Condition.IN);
+            soft.assertThat(value).isInstanceOf(ArrayQueryValue.class);
+            soft.assertThat(ArrayQueryValue.class.cast(value).get()).hasSize(2)
+                    .contains(EnumQueryValue.of(DayOfWeek.MONDAY), EnumQueryValue.of(DayOfWeek.SUNDAY));
 
         });
     }
 
+
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"WHERE  name NOT LIKE 'A%'", "FROM entity WHERE name NOT LIKE 'A%'"})
-    void shouldNegateLike(String query){
+    @ValueSource(strings = {"DELETE FROM entity WHERE age NOT IN (10, 20)"})
+    void shouldNegateBetween(String query){
         var deleteQuery = deleteProvider.apply(query);
 
         SoftAssertions.assertSoftly(soft -> {
@@ -84,14 +99,16 @@ class DeleteJakartaDataQueryProviderLikeTest {
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
             soft.assertThat(conditions).hasSize(1);
-            soft.assertThat(conditions.get(0).name()).isEqualTo("name");
-            soft.assertThat(conditions.get(0).value()).isEqualTo(StringQueryValue.of("A%"));
+            soft.assertThat(conditions.get(0).name()).isEqualTo("age");
+            soft.assertThat(conditions.get(0).value()).isInstanceOf(ArrayQueryValue.class);
+            soft.assertThat(ArrayQueryValue.class.cast(conditions.get(0).value()).get()).hasSize(2)
+                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
 
         });
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"WHERE  age = 10 AND name LIKE 'test'", "FROM entity WHERE  age = 10 AND name LIKE 'test'"})
+    @ValueSource(strings = {"DELETE FROM entity WHERE name = 'Otavio' AND age IN (10, 20)"})
     void shouldCombineAnd(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -106,16 +123,22 @@ class DeleteJakartaDataQueryProviderLikeTest {
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
             soft.assertThat(conditions).hasSize(2);
-            soft.assertThat(conditions.get(0).name()).isEqualTo("age");
-            soft.assertThat(conditions.get(0).value()).isEqualTo(NumberQueryValue.of(10));
 
-            soft.assertThat(conditions.get(1).name()).isEqualTo("name");
-            soft.assertThat(conditions.get(1).value()).isEqualTo(StringQueryValue.of("test"));
+            QueryCondition equalsCondition = conditions.get(0);
+            soft.assertThat(equalsCondition.name()).isEqualTo("name");
+            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
+
+            QueryCondition betweenCondition = conditions.get(1);
+            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.IN);
+            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
+            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
+                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
+
         });
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"WHERE name LIKE 'test' AND age = 10", "FROM entity WHERE name LIKE 'test' AND age = 10"})
+    @ValueSource(strings = "DELETE FROM entity WHERE age IN (10, 20) AND name = 'Otavio'")
     void shouldCombineAnd2(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -129,18 +152,21 @@ class DeleteJakartaDataQueryProviderLikeTest {
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
-            soft.assertThat(conditions).hasSize(2);
-            soft.assertThat(conditions.get(1).name()).isEqualTo("age");
-            soft.assertThat(conditions.get(1).value()).isEqualTo(NumberQueryValue.of(10));
+            QueryCondition equalsCondition = conditions.get(1);
+            soft.assertThat(equalsCondition.name()).isEqualTo("name");
+            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
 
-            soft.assertThat(conditions.get(0).name()).isEqualTo("name");
-            soft.assertThat(conditions.get(0).value()).isEqualTo(StringQueryValue.of("test"));
+            QueryCondition betweenCondition = conditions.get(0);
+            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.IN);
+            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
+            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
+                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
         });
     }
 
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"WHERE  age = 10 OR name LIKE 'test'", "FROM entity WHERE  age = 10 OR name LIKE 'test'"})
+    @ValueSource(strings = {"DELETE FROM entity WHERE  name = 'Otavio' OR name IN (10, 20)"})
     void shouldCombineOr(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -154,17 +180,20 @@ class DeleteJakartaDataQueryProviderLikeTest {
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
-            soft.assertThat(conditions).hasSize(2);
-            soft.assertThat(conditions.get(0).name()).isEqualTo("age");
-            soft.assertThat(conditions.get(0).value()).isEqualTo(NumberQueryValue.of(10));
+            QueryCondition equalsCondition = conditions.get(0);
+            soft.assertThat(equalsCondition.name()).isEqualTo("name");
+            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
 
-            soft.assertThat(conditions.get(1).name()).isEqualTo("name");
-            soft.assertThat(conditions.get(1).value()).isEqualTo(StringQueryValue.of("test"));
+            QueryCondition betweenCondition = conditions.get(1);
+            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.IN);
+            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
+            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
+                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
         });
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"WHERE name LIKE 'test' OR age = 10", "FROM entity WHERE name LIKE 'test' OR age = 10"})
+    @ValueSource(strings = "DELETE FROM entity WHERE age IN (10, 20) OR name = 'Otavio'")
     void shouldCombineOr2(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -178,12 +207,15 @@ class DeleteJakartaDataQueryProviderLikeTest {
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
-            soft.assertThat(conditions).hasSize(2);
-            soft.assertThat(conditions.get(1).name()).isEqualTo("age");
-            soft.assertThat(conditions.get(1).value()).isEqualTo(NumberQueryValue.of(10));
+            QueryCondition equalsCondition = conditions.get(1);
+            soft.assertThat(equalsCondition.name()).isEqualTo("name");
+            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
 
-            soft.assertThat(conditions.get(0).name()).isEqualTo("name");
-            soft.assertThat(conditions.get(0).value()).isEqualTo(StringQueryValue.of("test"));
+            QueryCondition betweenCondition = conditions.get(0);
+            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.IN);
+            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
+            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
+                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
         });
     }
