+++ jnosql-communication/jnosql-communication-query/src/test/java/org/eclipse/jnosql/communication/query/data/DeleteJakartaDataQueryProviderLikeTest.java
@@ -9,14 +9,12 @@
  *  Contributors:
  *  Otavio Santana
  */
-package org.eclipse.jnosql.communication.query.data.delete;
+package org.eclipse.jnosql.communication.query.data;
 
 import org.assertj.core.api.SoftAssertions;
 import org.eclipse.jnosql.communication.Condition;
-import org.eclipse.jnosql.communication.query.ArrayQueryValue;
 import org.eclipse.jnosql.communication.query.ConditionQueryValue;
 import org.eclipse.jnosql.communication.query.NumberQueryValue;
-import org.eclipse.jnosql.communication.query.QueryCondition;
 import org.eclipse.jnosql.communication.query.QueryValue;
 import org.eclipse.jnosql.communication.query.StringQueryValue;
 import org.eclipse.jnosql.communication.query.data.DeleteProvider;
@@ -24,7 +22,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class DeleteJakartaDataQueryProviderBetweenTest {
+class DeleteJakartaDataQueryProviderLikeTest {
 
 
     private DeleteProvider deleteProvider;
@@ -35,30 +33,44 @@ class DeleteJakartaDataQueryProviderBetweenTest {
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"DELETE FROM entity WHERE age BETWEEN 10 AND 20"})
-    void shouldBetween(String query){
-        var selectQuery = deleteProvider.apply(query);
+    @ValueSource(strings = "DELETE FROM entity WHERE  name LIKE 'A%'")
+    void shouldLike(String query){
+        var deleteQuery = deleteProvider.apply(query);
 
         SoftAssertions.assertSoftly(soft -> {
-            soft.assertThat(selectQuery.fields()).isEmpty();
-            soft.assertThat(selectQuery.entity()).isEqualTo("entity");
-            soft.assertThat(selectQuery.where()).isNotEmpty();
-            var where = selectQuery.where().orElseThrow();
+            soft.assertThat(deleteQuery.fields()).isEmpty();
+            soft.assertThat(deleteQuery.entity()).isEqualTo("entity");
+            soft.assertThat(deleteQuery.where()).isNotEmpty();
+            var where = deleteQuery.where().orElseThrow();
             var condition = where.condition();
             QueryValue<?> value = condition.value();
-            soft.assertThat(condition.name()).isEqualTo("age");
-            soft.assertThat(condition.condition()).isEqualTo(Condition.BETWEEN);
-            soft.assertThat(value).isInstanceOf(ArrayQueryValue.class);
-            soft.assertThat(ArrayQueryValue.class.cast(value).get()).hasSize(2)
-                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
+            soft.assertThat(condition.condition()).isEqualTo(Condition.LIKE);
+            soft.assertThat(value).isEqualTo(StringQueryValue.of("A%"));
 
         });
     }
 
+    @ParameterizedTest(name = "Should parser the query {0}")
+    @ValueSource(strings = "DELETE FROM entity WHERE  name LIKE \"A%\"")
+    void shouldLikeDoubleQuote(String query){
+        var deleteQuery = deleteProvider.apply(query);
+
+        SoftAssertions.assertSoftly(soft -> {
+            soft.assertThat(deleteQuery.fields()).isEmpty();
+            soft.assertThat(deleteQuery.entity()).isEqualTo("entity");
+            soft.assertThat(deleteQuery.where()).isNotEmpty();
+            var where = deleteQuery.where().orElseThrow();
+            var condition = where.condition();
+            QueryValue<?> value = condition.value();
+            soft.assertThat(condition.condition()).isEqualTo(Condition.LIKE);
+            soft.assertThat(value).isEqualTo(StringQueryValue.of("A%"));
+
+        });
+    }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"DELETE FROM entity WHERE age NOT BETWEEN 10 AND 20"})
-    void shouldNegateBetween(String query){
+    @ValueSource(strings = "DELETE FROM entity WHERE  name NOT LIKE 'A%'")
+    void shouldNegateLike(String query){
         var deleteQuery = deleteProvider.apply(query);
 
         SoftAssertions.assertSoftly(soft -> {
@@ -72,16 +84,14 @@ class DeleteJakartaDataQueryProviderBetweenTest {
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
             soft.assertThat(conditions).hasSize(1);
-            soft.assertThat(conditions.get(0).name()).isEqualTo("age");
-            soft.assertThat(conditions.get(0).value()).isInstanceOf(ArrayQueryValue.class);
-            soft.assertThat(ArrayQueryValue.class.cast(conditions.get(0).value()).get()).hasSize(2)
-                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
+            soft.assertThat(conditions.get(0).name()).isEqualTo("name");
+            soft.assertThat(conditions.get(0).value()).isEqualTo(StringQueryValue.of("A%"));
 
         });
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"DELETE FROM entity WHERE name = 'Otavio' AND age BETWEEN 10 AND 20"})
+    @ValueSource(strings = "DELETE FROM entity WHERE  age = 10 AND name LIKE 'test'")
     void shouldCombineAnd(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -96,22 +106,16 @@ class DeleteJakartaDataQueryProviderBetweenTest {
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
             soft.assertThat(conditions).hasSize(2);
+            soft.assertThat(conditions.get(0).name()).isEqualTo("age");
+            soft.assertThat(conditions.get(0).value()).isEqualTo(NumberQueryValue.of(10));
 
-            QueryCondition equalsCondition = conditions.get(0);
-            soft.assertThat(equalsCondition.name()).isEqualTo("name");
-            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
-
-            QueryCondition betweenCondition = conditions.get(1);
-            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.BETWEEN);
-            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
-            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
-                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
-
+            soft.assertThat(conditions.get(1).name()).isEqualTo("name");
+            soft.assertThat(conditions.get(1).value()).isEqualTo(StringQueryValue.of("test"));
         });
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"DELETE FROM entity WHERE age BETWEEN 10 AND 20 AND name = 'Otavio'"})
+    @ValueSource(strings = "DELETE FROM entity WHERE name LIKE 'test' AND age = 10")
     void shouldCombineAnd2(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -125,21 +129,18 @@ class DeleteJakartaDataQueryProviderBetweenTest {
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
-            QueryCondition equalsCondition = conditions.get(1);
-            soft.assertThat(equalsCondition.name()).isEqualTo("name");
-            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
+            soft.assertThat(conditions).hasSize(2);
+            soft.assertThat(conditions.get(1).name()).isEqualTo("age");
+            soft.assertThat(conditions.get(1).value()).isEqualTo(NumberQueryValue.of(10));
 
-            QueryCondition betweenCondition = conditions.get(0);
-            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.BETWEEN);
-            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
-            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
-                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
+            soft.assertThat(conditions.get(0).name()).isEqualTo("name");
+            soft.assertThat(conditions.get(0).value()).isEqualTo(StringQueryValue.of("test"));
         });
     }
 
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"DELETE FROM entity WHERE  name = 'Otavio' OR name BETWEEN 10 AND 20"})
+    @ValueSource(strings = "DELETE FROM entity WHERE  age = 10 OR name LIKE 'test'")
     void shouldCombineOr(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -153,20 +154,17 @@ class DeleteJakartaDataQueryProviderBetweenTest {
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
-            QueryCondition equalsCondition = conditions.get(0);
-            soft.assertThat(equalsCondition.name()).isEqualTo("name");
-            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
+            soft.assertThat(conditions).hasSize(2);
+            soft.assertThat(conditions.get(0).name()).isEqualTo("age");
+            soft.assertThat(conditions.get(0).value()).isEqualTo(NumberQueryValue.of(10));
 
-            QueryCondition betweenCondition = conditions.get(1);
-            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.BETWEEN);
-            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
-            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
-                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
+            soft.assertThat(conditions.get(1).name()).isEqualTo("name");
+            soft.assertThat(conditions.get(1).value()).isEqualTo(StringQueryValue.of("test"));
         });
     }
 
     @ParameterizedTest(name = "Should parser the query {0}")
-    @ValueSource(strings = {"DELETE FROM entity WHERE age BETWEEN 10 AND 20 OR name = 'Otavio'"})
+    @ValueSource(strings = "DELETE FROM entity WHERE name LIKE 'test' OR age = 10")
     void shouldCombineOr2(String query){
         var deleteQuery = deleteProvider.apply(query);
 
@@ -180,15 +178,12 @@ class DeleteJakartaDataQueryProviderBetweenTest {
 
             var values = (ConditionQueryValue) condition.value();
             var conditions = values.get();
-            QueryCondition equalsCondition = conditions.get(1);
-            soft.assertThat(equalsCondition.name()).isEqualTo("name");
-            soft.assertThat(equalsCondition.value()).isEqualTo(StringQueryValue.of("Otavio"));
+            soft.assertThat(conditions).hasSize(2);
+            soft.assertThat(conditions.get(1).name()).isEqualTo("age");
+            soft.assertThat(conditions.get(1).value()).isEqualTo(NumberQueryValue.of(10));
 
-            QueryCondition betweenCondition = conditions.get(0);
-            soft.assertThat(betweenCondition.condition()).isEqualTo(Condition.BETWEEN);
-            soft.assertThat(betweenCondition.value()).isInstanceOf(ArrayQueryValue.class);
-            soft.assertThat(ArrayQueryValue.class.cast(betweenCondition.value()).get()).hasSize(2)
-                    .contains(NumberQueryValue.of(10), NumberQueryValue.of(20));
+            soft.assertThat(conditions.get(0).name()).isEqualTo("name");
+            soft.assertThat(conditions.get(0).value()).isEqualTo(StringQueryValue.of("test"));
         });
     }
